backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-genai==0.8.0  # Updated to the new Google Gen AI SDK
python-docx==1.1.2
pydantic>=2.7.4
pydantic-settings==2.2.1  # Added for BaseSettings support in Pydantic v2
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
python-multipart==0.0.9  # For handling file uploads
# sqlite3 is part of the standard library
sqlalchemy-utils
alembic==1.12.1  # For database migrations
google-generativeai>=0.3.0
psycopg2-binary==2.9.9  # For PostgreSQL support

# RAG Dependencies
langchain-core>=0.3.63
langchain-openai>=0.3.18
langchain_text_splitters>=0.0.1
langgraph>=0.4.7
tiktoken>=0.9.0
langdetect>=1.0.9
qdrant-client>=1.14.2
langchain-qdrant>=0.2.0
openai>=1.8.0